btcusd-trader/requirements.txt
# Data & Analysis
pandas==2.2.0
numpy==1.24.3
scipy==1.12.0
pyarrow==15.0.0

# API & Networking
aiohttp==3.9.1
requests==2.31.0
python-binance==1.0.17

# Configuration
pyyaml==6.0.1

# Logging & Monitoring
python-telegram-bot==20.1

# Database
sqlite3  # Built-in with Python

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.22.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1

# Performance
numba==0.58.1  # Optional: for faster numpy operations
